first stage of refactor
