+++ .gitignore
@@ -8,6 +8,7 @@ test_data/
 *.ipynb*
 *.tif
 *.
+*.csv
 
 # dependencies
 node_modules
@@ -15,6 +16,7 @@ node_modules
 .pnp.js
 
 # testing
+test/
 coverage
 
 # production
