+++ backend/setup.sh
@@ -1,0 +1,4 @@
+
+#!/bin/zsh
+conda activate array_tomography_tool
+export PYTHONPATH="$PYTHONPATH:/Users/Lewis/cernbox/Personal/ArrayTomographyTool/backend"
